requirements.txt
xgboost==1.6.2
pandas
scikit-learn==1.1.2
imblearn==0.0
scipy==1.9.3
numpy==1.21.2
catboost==1.1.1
seaborn==0.12.2
kneed==0.8.1
prettytable==3.5.0
matplotlib==3.6.2
miceforest==5.6.3
pymongo[srv]==4.2.0
from_root==1.0.2
pyyaml==5.4.1
fastapi==0.78.0
uvicorn==0.18.3
python-multipart==0.0.5
-e .